requirements.txt
fastapi==0.62.0
requests==2.20.0
uvicorn==0.13.1
starlette==0.13.6
python-dotenv==0.19.0
pymongo
python-decouple
